rename funcx_ws to funcx_websocket_service
