+++ Dockerfile
@@ -11,7 +11,7 @@ RUN pip install -r requirements.txt
 # RUN pip install --disable-pip-version-check uwsgi
 
 # COPY uwsgi.ini .
-COPY ./funcx_ws/ ./funcx_ws/
+COPY ./funcx_websocket_service/ ./funcx_websocket_service/
 COPY ./run.py .
 
 USER uwsgi
