+++ funcx_websocket_service/auth.py
